Add support for difference between abstract and detail
For games that start at around 9:00 JST every day, it often happens that
the time recorded in abstract is just before 9:00 JST and the time
recorded in detail is just after 9:00 JST, even though the UUID of the
game is the same.  As a result, the S3 key prefixes of abstract and
detail of a game with the same UUID may be different, i.e., the S3 key
prefix of detail may be one day after the key prefix of abstract.  This
commit adds support for such situation.